Essentials/chat-application/server/requirements.txt
python-socketio
werkzeug